requirements.txt
Appium-Python-Client==0.40
certifi==2019.3.9
chardet==3.0.4
idna==2.8
PyMySQL==0.9.3
redis==3.2.1
requests==2.21.0
selenium==3.141.0
urllib3==1.24.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
